+++ packages/cloud_firestore/lib/src/query.dart
@@ -50,9 +50,8 @@ class Query {
 
   /// Notifies of query results at this location
   // TODO(jackson): Reduce code duplication with [DocumentReference]
-  Stream<QuerySnapshot> snapshots(
-      {MetadataChanges metadataChanges = MetadataChanges.exclude}) {
-    assert(metadataChanges != null);
+  Stream<QuerySnapshot> snapshots({bool includeMetadataChanges = false}) {
+    assert(includeMetadataChanges != null);
     Future<int> _handle;
     // It's fine to let the StreamController be garbage collected once all the
     // subscribers have cancelled; this analyzer warning is safe to ignore.
@@ -66,7 +65,7 @@ class Query {
             'path': _path,
             'isCollectionGroup': _isCollectionGroup,
             'parameters': _parameters,
-            'metadataChanges': _getMetadataChangesString(metadataChanges),
+            'includeMetadataChanges': includeMetadataChanges,
           },
         ).then<int>((dynamic result) => result);
         _handle.then((int handle) {
@@ -76,7 +75,7 @@ class Query {
       onCancel: () {
         _handle.then((int handle) async {
           await Firestore.channel.invokeMethod<void>(
-            'Query#removeListener',
+            'removeListener',
             <String, dynamic>{'handle': handle},
           );
           Firestore._queryObservers.remove(handle);